symbol link i2c-suniv.c to build dir
